99 Bottles of Beer
"99 Bottles of Beer" is an anonymous folk song dating to the mid-20th century. It is a traditional reverse counting song in both the United States and Canada. It is popular to sing on road trips, as it has a very repetitive format which is easy to memorize and can take a long time when families sing.